multi: move onion messenger into module
